+++ scripts/completions/skifflm.zsh
@@ -71,7 +71,7 @@ _skifflm() {
         '--verbose[Verbose logs]'
         '--debug[Debug mode]'
     )
-    _arguments "${opts[@]}" '1:subcommand:(run model chat-template config server)' '*:model file:_files'
+    _arguments "${opts[@]}" '1:subcommand:(run model chat-template openai config server)' '*:model file:_files'
 }
 
 _skifflm "$@"